requirements.txt
fastapi
uvicorn
google-generativeai
firebase-admin
httpx
streamlit
pandas
python-multipart
